+++ main/src/main/java/net/oneandone/stool/client/cli/Globals.java
@@ -16,65 +16,36 @@
 package net.oneandone.stool.client.cli;
 
 import net.oneandone.inline.Console;
-import net.oneandone.stool.server.util.Environment;
-import net.oneandone.stool.server.util.LogOutputStream;
-import net.oneandone.stool.server.util.Logging;
 import net.oneandone.stool.server.util.Server;
 import net.oneandone.sushi.fs.World;
 import net.oneandone.sushi.fs.file.FileNode;
-import net.oneandone.sushi.io.InputLogStream;
-import net.oneandone.sushi.io.MultiOutputStream;
-import net.oneandone.sushi.util.Separator;
 
 import java.io.IOException;
-import java.io.OutputStream;
+import java.io.PrintWriter;
 import java.lang.reflect.InvocationTargetException;
 
 /** Basically a session factory */
 public class Globals {
     public static Globals create(World world, FileNode itHome, String[] args) throws IOException {
-        Environment environment;
-        FileNode home;
-        Logging logging;
-        FileNode tmp;
         Console console;
+        PrintWriter itOut;
 
-        environment = Environment.loadSystem();
         if (itHome != null) {
-            environment.setHome(itHome);
-        }
-        home = environment.locateHome(world);
-        if (home.exists()) {
-            logging = Logging.forHome(home, environment.detectUser());
+            itOut = new PrintWriter(itHome.join("client-output").newAppender(), true);
+            console = new Console(itOut, itOut, System.in);
         } else {
-            tmp = world.getTemp().createTempDirectory();
-            logging = new Logging("1", tmp.join("homeless"), environment.detectUser());
+            console = Console.create();
         }
-        if (itHome != null) {
-            OutputStream devNull = MultiOutputStream.createNullStream();
-            console = console(logging, devNull, devNull);
-        } else {
-            console = console(logging, System.out, System.err);
-        }
-        logging.command("stool " + Separator.SPACE.join(args));
-        return new Globals(itHome, logging, args, console, world);
+        return new Globals(itHome, args, console, world);
     }
-
-    public static Console console(Logging logging, OutputStream out, OutputStream err) {
-        return new Console(logging.writer(out, "OUT"), logging.writer(err, "ERR"),
-                new InputLogStream(System.in, new LogOutputStream(logging, "IN")));
-    }
-
 
     private final FileNode itHome;
-    public final Logging logging;
     private final String[] args;
     public final Console console;
     public final World world;
 
-    public Globals(FileNode itHome, Logging logging, String[] args, Console console, World world) {
+    public Globals(FileNode itHome, String[] args, Console console, World world) {
         this.itHome = itHome;
-        this.logging = logging;
         this.args = args;
         this.console = console;
         this.world = world;